Move TextPiece, CodePiece, CommentPiece, and SpacePiece to another file.
piece.dart was getting a little unwieldy. This also deliberately makes
calculateContainsNewline() and calculateTotalCharacters() public to
pave the way for other Piece classes overriding them.
